move retrieving current temp to useTemperature helper
