+++ src/components/Temperature.tsx
@@ -1,12 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
 import { Tooltip, Typography } from '@material-ui/core';
 import { ErrorOutline as ErrorOutlineIcon } from '@material-ui/icons';
 import { Skeleton } from '@material-ui/lab';
-import axios from 'axios';
 import styled from 'styled-components';
 
 import { colors, font } from '../App.styles';
+import useTemperature from './useTemperature';
 
 const StyledContainer = styled.aside`
   color: ${colors.white};
@@ -50,80 +50,18 @@ const StyledTooltip = styled(props => (
   }
 ` as typeof Tooltip;
 
-interface TCoordMeta {
-  data: {
-    properties: {
-      radarStation: string;
-    };
-  };
-}
-
-interface TLatestObservation {
-  data: {
-    properties: {
-      temperature: {
-        value: number;
-      };
-    };
-  };
-}
-
-const convertCelsiusToFarenheit = (cTemp: number) => cTemp * (9 / 5) + 32;
-
 const Temperature = () => {
-  const [temperature, setTemperature] = useState<undefined | null | number>();
-
-  const boulderCoords = {
-    lat: 40.015,
-    lng: -105.2705
-  };
-
-  useEffect(() => {
-    const getCoordMeta = (): Promise<TCoordMeta> =>
-      axios.get(
-        `https://api.weather.gov/points/${boulderCoords.lat},${
-          boulderCoords.lng
-        }`
-      );
-    const getLatestObservation = (
-      stationId: string
-    ): Promise<TLatestObservation> =>
-      axios.get(
-        `https://api.weather.gov/stations/${stationId}/observations/latest`,
-        {}
-      );
-
-    const getTemperature = async () => {
-      try {
-        const { data: coordMeta } = await getCoordMeta();
-        const { data: latestObservation } = await getLatestObservation(
-          coordMeta.properties.radarStation
-        );
-        const currentTempCelsius =
-          latestObservation.properties.temperature.value;
-        const currentTempFarenheit = convertCelsiusToFarenheit(
-          currentTempCelsius
-        );
-        const roundedTemp = Math.round(currentTempFarenheit);
-        setTemperature(roundedTemp);
-      } catch (e) {
-        setTemperature(null);
-      }
-    };
-
-    getTemperature();
-  }, [setTemperature]);
-
+  const temp = useTemperature();
   return (
     <StyledContainer>
       <Typography variant="srOnly" component="h2">
         Current Temperature
       </Typography>
       <StyledTemp>
-        {temperature === undefined && (
+        {temp === undefined && (
           <Skeleton variant="rect" height={19} width={40} />
         )}
-        {temperature === null && (
+        {temp === null && (
           <StyledTooltip
             placement="left"
             title="temperature currently unavailable"
@@ -131,7 +69,7 @@ const Temperature = () => {
             <ErrorOutlineIcon fontSize="small" tabIndex={0} />
           </StyledTooltip>
         )}
-        {temperature && <StyledDegrees>{temperature}&deg;</StyledDegrees>}
+        {temp && <StyledDegrees>{temp}&deg;</StyledDegrees>}
         <Typography variant="srOnly">in</Typography>
         <StyledLocation>Boulder, CO</StyledLocation>
       </StyledTemp>
